Added some context for wheel constants in Constants.java.
We will be using a shorter drivetrain width on the actual robot so the lifts don't push us over the 18in size limit. The part that will allow us to make the robot smaller won't actually be in until later this week though, so we will be testing with a wider robot and will need to use the 10 hole crossbar values. This should be fine, because our robot code--if it is written correctly--should work with any size robot so long as we measure the wheel distances correctly.