We need to test this but if it works, we have field coordinates setup now.
The way it works:

(0,0,0) in field coordinates is in the center of the field facing away from the red alliance, with a positive x moving right, a positive y moving forward, and a positive rotation rotating counterclockwise.

This will work if we give intialPose in the constructor in field coordinates.